Scan strings in the diag_table file for unallowed characters
It was found that if a user misplaces a quotation mark (") in one
of the strings in the diag_table then commas, or other characters
could become part of the string causing strange errors either during
the run, or during post processing of the data.  This fix scans all
strings in the diag_table for un-allowed characters, and thorws a
WARNING if found.