xgrid.f90 update; restore fms_io_nml flags
The alltoall-based exchange grid calculation is restored in this
commit.  It is based on the development version of FMS, since it has not
yet been integrated into the Ulm release.  It is based on the current
Alltoall implementation in the repo, with some minor changes by Zhi
Liang.

The following deprecated FMS IO namelist settings have also been
restored:

* ``threading_write``
* ``fileset_write``