OBC update for new FMS IO: boundary communicators
The new FMS IO only allows one rank to open a file, and then
`MPI_BCAST`s the data to the other ranks.  This was causing model hangs
on the OBC code (via `init_external_field`), where only ranks along the
boundaries would open a file.

This was resolved by switching to the local boundary communicator during
the initialization IO for the barotropic and tracer initialization.

This required some copypasta from `ocean_obc_barotropic.F90` to
`ocean_obc.F90`, since only the barotropic part was properly
constructing the boundary communicators.  This should probably be
consolidated someday in the future, since there are now identical
duplicates of these communicators (along with the code duplication).