Convert data source paths to URLs
In `data_sources.csv`, some target data files are preceded bya URL
(file://) while others are simply filepaths.  This appears to cause
problems when wget tries to read them, creating empty files with the
same name.

This patch forces local paths to be redefined as URLs if they exist.
Not fool-proof, but a slight improvement.